requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
binaryornot==0.4.4
celery==5.4.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cron-descriptor==1.4.3
cssselect2==0.7.0
dj-database-url==2.2.0
Django==5.0.6
django-adminlte2==0.4.3
django-background-tasks==1.2.8
django-browser-reload==1.13.0
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-filer==3.1.3
django-humanize==0.1.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-polymorphic==3.1.0
django-tailwind==3.8.0
django-timezone-field==7.0
django-tinymce==4.1.0
easy-thumbnails==2.8.5
facebook-sdk==3.1.0
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
humanize==4.9.0
idna==3.7
Jinja2==3.1.4
kombu==5.4.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
tweepy==4.14.0
twilio==9.2.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.7.0
yarl==1.9.4
